requirements.txt
# 美妆内容标签抽取系统依赖包

# 核心依赖
pydantic>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
PyYAML>=6.0
openpyxl>=3.1.0

# 文本处理
jieba>=0.42.1
zhconv>=1.4.3
# re2>=1.0.0  # 可选依赖，如果需要可以单独安装
rapidfuzz>=3.0.0

# LLM客户端
requests>=2.28.0
openai>=1.0.0
dashscope>=1.14.0
aiohttp>=3.8.0

# 数据导出
pyarrow>=10.0.0
fastparquet>=0.8.0

# 日志和配置
loguru>=0.7.0
click>=8.0.0

# 开发和测试
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# 可选依赖（用于特定功能）
# 如果需要更好的中文分词
# pkuseg>=0.0.25
# LAC>=2.1.0

# 如果需要GPU加速的文本处理
# torch>=1.13.0
# transformers>=4.20.0

# 如果需要更多的数据格式支持
# xlsxwriter>=3.0.0
# h5py>=3.7.0